+++ NEWS.md
@@ -9,7 +9,7 @@
 * Added DSD_ReadStream as an alias for DSD_ReadCSV.
 * Added DSD_Combination to combine streams.
 * stream now implements a standard predict function (get_assignment() is now deprecated).
-* update now returns information like cluster assignments as a data.frame.
+* update() now returns information like cluster assignments as a data.frame.
 * DSC implementation now have a formula argument to decide what variables should be used for 
   clustering.